backend/requirements.txt
Flask==3.1.0
flask-cors==4.0.0
google-generativeai==0.5.0